[ai-0.36-63] Core/Playerbots: Fixing DB SQL query issues. Thanks to Kittnz. [trinity only]

+++ src/plugins/ahbot/AhBot.cpp
@@ -886,7 +886,7 @@ void AhBot::LoadRandomBots()
         {
             Field* fields = result->Fetch();
             uint32 guid = fields[0].GetUInt32();
-            uint32 race = fields[1].GetUInt32();
+            uint8 race = fields[1].GetUInt8();
             uint32 auctionHouse = PlayerbotAI::IsOpposing(race, RACE_HUMAN) ? 2 : 1;
             bidders[auctionHouse].push_back(guid);
             bidders[3].push_back(guid);
