+++ src/plugins/playerbot/RandomPlayerbotMgr.h
@@ -58,9 +58,9 @@ class RandomPlayerbotMgr : public PlayerbotHolder
         uint32 AddRandomBot(bool alliance);
         bool ProcessBot(uint32 bot);
         void ScheduleRandomize(uint32 bot, uint32 time);
-        void RandomTeleport(Player* bot, uint32 mapId, float teleX, float teleY, float teleZ);
+        void RandomTeleport(Player* bot, uint16 mapId, float teleX, float teleY, float teleZ);
         void RandomTeleport(Player* bot, vector<WorldLocation> &locs);
-        uint32 GetZoneLevel(uint32 mapId, float teleX, float teleY, float teleZ);
+        uint32 GetZoneLevel(uint16 mapId, float teleX, float teleY, float teleZ);
 
     private:
         vector<Player*> players;
